Add files via upload
This code compute the orbit by monodromy action on vanishing cycles in mirror quintic Calabi-Yau threefolds. The global monodromy for a family of mirror quintic  CY threefolds is represented by matrices M0 and M1, also there are two vectors which represent a Lagrangian sphere and a Lagrangian torus. For prime numbers p, we can compute the orbit of these Lagrangians submanifolds, in the homology with coefficients in Zp (see arXiv:2005.07736).